Vetical UI - Expandable and collapsing Tabs working
The Tabs expand and collapse but they can open one after another and then it gets very overwhelming. I also tried to add a scroll view but I have to learn how to separate the screen in two because now the mouse scroll is used for zoom in/out for the 3D model.